+++ python/arki-scan/dispatch.cc
@@ -59,16 +59,18 @@ DispatchResults MetadataDispatch::process(dataset::Reader& ds,
     // Read
     try
     {
-        ds.query_data(Matcher(), [&](std::shared_ptr<Metadata> md) {
-            auto scanner = scan::Scanner::get_scanner(md->source().format);
-            scanner->normalize_before_dispatch(*md);
-            // TODO: preprocess here, leave untouched or return inline
-            partial_batch_data_size += md->data_size();
-            partial_batch->acquire(move(md));
-            if (flush_threshold && partial_batch_data_size > flush_threshold)
-                process_partial_batch(name, stats);
-            return true;
-        });
+        ds.query_data(
+            query::Data(Matcher(), true), [&](std::shared_ptr<Metadata> md) {
+                auto scanner = scan::Scanner::get_scanner(md->source().format);
+                scanner->normalize_before_dispatch(*md);
+                // TODO: preprocess here, leave untouched or return inline
+                partial_batch_data_size += md->data_size();
+                partial_batch->acquire(move(md));
+                if (flush_threshold &&
+                    partial_batch_data_size > flush_threshold)
+                    process_partial_batch(name, stats);
+                return true;
+            });
         if (!partial_batch->empty())
             process_partial_batch(name, stats);
     }